부하 테스트 결과 추가 및 버그 수정

+++ src/main/java/com/astar/gostudy_be/domain/user/service/UserService.java
@@ -80,7 +80,7 @@ public class UserService implements UserDetailsService {
     public Long updateProfile(ProfileUpdateDto profileUpdateDto, Account account) {
         String filename = null;
         if(profileUpdateDto.getImage() != null) {
-            if (!Objects.equals(account.getImage(), "default.png") || !account.getImage().contains("http")) {
+            if (!Objects.equals(account.getImage(), "default.png") && !account.getImage().contains("http")) {
                 new File(Config.UPLOAD_FILE_PATH + "uploads/profile/images/" + account.getImage()).delete();
                 new File(Config.UPLOAD_FILE_PATH + "uploads/profile/thumbnail_images/thumbnail_" + account.getImage()).delete();
             }
